api calls away from scene logic
• Перенос логики вызова списка авторов из реализации сцены в отдельный файл

+++ src/scenes/authorSearch.ts
@@ -2,9 +2,9 @@
 /* eslint-disable import/extensions */
 
 import { Markup, Scenes } from 'telegraf';
-import axios from 'axios';
 import Bot from '../types/bot';
 import { author } from '../interfaces/baseObj';
+import API from '../api';
 
 export default class authorSearch {
   static list: Array<author>
@@ -38,18 +38,25 @@ export default class authorSearch {
 
     scene.enter(async (ctx) => {
       try {
-        authorSearch.list = (await axios.get(`http://localhost:3000/authors?country=${ctx.session.countryId}`)).data;
+        this.list = await API.getAuthors(ctx.session.countryId);
       } catch (error) {
         ctx.reply('Что-то не так с базой данных');
         ctx.scene.enter('start');
         throw new Error(error);
       }
 
-      authorSearch.hearings(scene);
+      if (typeof this.list[0] === 'boolean') {
+        ctx.reply('Что-то не так с базой данных');
+        ctx.scene.enter('start');
+      }
 
-      await ctx.deleteMessage(ctx.session.prevMessage);
+      if (typeof this.list === 'object') {
+        authorSearch.hearings(scene);
 
-      ctx.session.prevMessage = (await ctx.reply('Выберите автора из меню', Markup.keyboard(authorSearch.authorButtons()).oneTime())).message_id;
+        await ctx.deleteMessage(ctx.session.prevMessage);
+
+        ctx.session.prevMessage = (await ctx.reply('Выберите автора из меню', Markup.keyboard(authorSearch.authorButtons()).oneTime())).message_id;
+      }
     });
 
     return scene;
